requirements.txt
matplotlib
nltk
pandas
progressbar2
pywikibot
pyyaml
requests
seaborn
sidetable
sparqlwrapper
torch
wiki-dump-reader
wordcloud
